bugfix printf and float

+++ examples/osd/arduino-climate3/resources/res-htu21dhum.c
@@ -64,12 +64,12 @@ res_get_handler(void *request, void *response, uint8_t *buffer, uint16_t preferr
 
   if(accept == -1 || accept == REST.type.TEXT_PLAIN) {
     REST.set_header_content_type(response, REST.type.TEXT_PLAIN);
-    snprintf((char *)buffer, REST_MAX_CHUNK_SIZE, "%f", htu21d_hum);
+    snprintf((char *)buffer, REST_MAX_CHUNK_SIZE, "%f", (double) htu21d_hum);
 
     REST.set_response_payload(response, buffer, strlen((char *)buffer));
   } else if(accept == REST.type.APPLICATION_JSON) {
     REST.set_header_content_type(response, REST.type.APPLICATION_JSON);
-    snprintf((char *)buffer, REST_MAX_CHUNK_SIZE, "{'moisture':%f}", htu21d_hum);
+    snprintf((char *)buffer, REST_MAX_CHUNK_SIZE, "{'moisture':%f}", (double) htu21d_hum);
 
     REST.set_response_payload(response, buffer, strlen((char *)buffer));
   } else {
